backend/pyproject.toml
[project]
name = "compass-backend"
version = "0.1.0"
description = "CompassCHW Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.1.0",
    "python-jose[cryptography]>=3.4.0",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "boto3>=1.36.0",
    "python-multipart>=0.0.19",
    "httpx>=0.28.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "S", "B"]
ignore = [
    "B008",   # FastAPI `Depends()` as default argument is the idiomatic pattern
    "E402",   # Lazy imports inside functions are used to avoid circular imports
    "UP042",  # StrEnum is 3.11+ but we use compatible `str, Enum` pattern
    "E501",   # Long lines in HTML templates are acceptable
    "S101",   # Assertions allowed in tests
    "E712",   # SQLAlchemy requires `== True/False` for column comparisons
    "N806",   # SQLAlchemy aliased() naming convention is PascalCase
    "F403",   # Wildcard imports used intentionally in models/__init__ for enum re-export
    "S105",   # "bearer" is not a password — false positive on OAuth2 token_type
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S105", "S106"]  # Hardcoded test credentials are fine in tests

[tool.pytest.ini_options]
asyncio_mode = "auto"
